+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 5.7.0
 
+- tidy3d mode solver accepts ncore and nclad floats. add file cache to tidy3d. thanks to `gt.modes.find_modes`
 - fix get_bundle [issue](https://github.com/gdsfactory/gdsfactory/issues/396)
 - clean cross-sections [PR](https://github.com/gdsfactory/gdsfactory/pull/398/files)
 - fix N/S routing in route_ports_to_side [PR](https://github.com/gdsfactory/gdsfactory/pull/395)
@@ -10,6 +11,8 @@
    - Use via_corner instead of wire_corner for bend function
    - Use MultiCrossSectionAngleSpec instead of CrossSectionSpec to define multiple cross sections
    - Avoids refactoring as much as possible so it doesn't interfere with current single-layer routing
+
+
 
 ## [5.6.12](https://github.com/gdsfactory/gdsfactory/pull/397)